Move sota.env from /usr/lib to /etc
The current way for the user to replace the default environment is by
copying the original aktualizr.service to /etc/systemd/system and
changing EnvironmentFile to point to a different file, which is not user
friendly.

Change the default environment path from /usr/lib/sota/sota.env to
/etc/sota/sota.env to allow user changes during runtime (e.g. pointing
out to a different server).

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>